pyproject.toml
[project]
name = "venue2playlist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "camhpj", email = "camhpjohnson@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "typer[all]>=0.9.0",
    "spotipy>=2.23.0",
    "httpx>=0.27.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "musicbrainzngs>=0.7.1",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "dateparser>=1.2.0",
    "structlog>=24.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
venue2playlist = "venue2playlist:main"

[build-system]
requires = ["uv_build>=0.8.24,<0.9.0"]
build-backend = "uv_build"
